Create README.md
I am continuously learning from my financial misses and hits that have broadly influenced the path I have taken to this point. I am an aspiring CYBER SECURITY AND FORENSICS Specialist driven by the desire to reduce or even eliminate humans dependance on Debt and over reliance on AID/HELP to bring the desired change in their lives. As this Journey STARTS, SET your eyes on the prize with me. WE WILL DO IT.